+++ lib/oli/activities/model/conditional_outcome.ex
@@ -26,12 +26,12 @@ defmodule Oli.Activities.Model.ConditionalOutcome do
     {:error, "invalid conditional outcome"}
   end
 
-  defp parse_action(%{"type" => "FeedbackAction"} = action),
+  defp parse_action(%{"type" => "FeedbackActionDesc"} = action),
     do: Oli.Activities.Model.FeedbackAction.parse(action)
 
-  defp parse_action(%{"type" => "NavigationAction"} = action),
+  defp parse_action(%{"type" => "NavigationActionDesc"} = action),
     do: Oli.Activities.Model.NavigationAction.parse(action)
 
-  defp parse_action(%{"type" => "StateUpdateAction"} = action),
+  defp parse_action(%{"type" => "StateUpdateActionDesc"} = action),
     do: Oli.Activities.Model.StateUpdateAction.parse(action)
 end