+++ lib/oli/activities/model/state_update_action.ex
@@ -2,10 +2,11 @@ defmodule Oli.Activities.Model.StateUpdateAction do
   defstruct [:id, :update]
 
   def parse(%{"id" => id, "update" => update}) do
-    %Oli.Activities.Model.StateUpdateAction{
-      id: id,
-      update: update
-    }
+    {:ok,
+     %Oli.Activities.Model.StateUpdateAction{
+       id: id,
+       update: update
+     }}
   end
 
   def parse(_) do