+++ lib/oli/delivery/sections/enrollment.ex
@@ -5,6 +5,7 @@ defmodule Oli.Delivery.Sections.Enrollment do
   schema "enrollments" do
     belongs_to :user, Oli.Accounts.User
     belongs_to :section, Oli.Delivery.Sections.Section
+    field :state, :map, default: %{}
 
     many_to_many :context_roles, Lti_1p3.DataProviders.EctoProvider.ContextRole,
       join_through: "enrollments_context_roles",
@@ -16,7 +17,7 @@ defmodule Oli.Delivery.Sections.Enrollment do
   @doc false
   def changeset(section, attrs) do
     section
-    |> cast(attrs, [:user_id, :section_id])
+    |> cast(attrs, [:user_id, :section_id, :state])
     |> validate_required([:user_id, :section_id])
   end
 end